Move Sequential stuff to own module
