+++ magic-wormhole.cabal
@@ -53,6 +53,7 @@ library
       MagicWormhole.Internal.Messages
       MagicWormhole.Internal.Peer
       MagicWormhole.Internal.Rendezvous
+      MagicWormhole.Internal.Sequential
       MagicWormhole.Internal.WebSockets
   default-language: Haskell2010
