+++ src/MagicWormhole/Internal/ApplicationProtocol.hs
@@ -19,11 +19,11 @@ import Control.Concurrent.STM.TVar
   , newTVar
   , readTVar
   )
-import qualified Data.HashMap.Lazy as HashMap
 
 import qualified MagicWormhole.Internal.ClientProtocol as ClientProtocol
 import qualified MagicWormhole.Internal.Messages as Messages
 import qualified MagicWormhole.Internal.Peer as Peer
+import qualified MagicWormhole.Internal.Sequential as Sequential
 
 -- | A Magic Wormhole peer-to-peer application session.
 --
@@ -39,7 +39,7 @@ data Session
   = Session
   { connection :: Peer.Connection
   , key :: ClientProtocol.SessionKey
-  , inbound :: Sequential Int (Messages.Phase, ClientProtocol.PlainText)
+  , inbound :: Sequential.Sequential Int (Messages.Phase, ClientProtocol.PlainText)
   , outbound :: TVar Int
   }
 
@@ -61,7 +61,7 @@ withSession conn sessionKey action = do
 
 -- | Construct a new session.
 newSession :: Peer.Connection -> ClientProtocol.SessionKey -> STM Session
-newSession conn sessionKey = Session conn sessionKey <$> sequenceBy getAppRank firstPhase <*> newTVar firstPhase
+newSession conn sessionKey = Session conn sessionKey <$> Sequential.sequenceBy getAppRank firstPhase <*> newTVar firstPhase
   where
     getAppRank (phase, _) =
       case phase of
@@ -93,7 +93,7 @@ runSession session action = do
       case msg of
         Left err -> pure $ Left err
         Right msg' -> do
-          atomically $ insert (inbound session) msg'
+          atomically $ Sequential.insert (inbound session) msg'
           readLoop
 
 -- | Send an encrypted message to the peer.
@@ -116,51 +116,4 @@ sendMessage session body = do
 --
 -- Obtain a 'Session' with 'withSession'.
 receiveMessage :: Session -> STM ClientProtocol.PlainText
-receiveMessage session = snd <$> next (inbound session)
-
--- | A thing that gets numbered items and then buffers them so you can read
--- them in order without gaps.
---
--- "counter" is an 'Enum' that is used to determine the ordering of the
--- elements, and "item" is the type of the items themselves.
-data Sequential counter item
-  = Sequential
-  { -- | The current value of the counter. The next item must have this rank.
-    current :: TVar counter
-  , -- | Where we store the items that have arrived out of order. A priority
-    -- queue might be a better choice.
-    buffer :: TVar (HashMap.HashMap counter item)
-  , -- | How to rank items. Each item has a "rank", which is an 'Enum'
-    -- (specifically, a "counter"). Items are yielded in order, without gaps,
-    -- by 'next'.
-    rank :: item -> counter
-  }
-
--- | Create a 'Sequential' value for a series of item.
-sequenceBy
-  :: (Hashable counter, Eq counter)
-  => (a -> counter) -- ^ How to rank items
-  -> counter -- ^ The expected rank of the first item
-  -> STM (Sequential counter a)
-sequenceBy rank' initial = Sequential <$> newTVar initial <*> newTVar mempty <*> pure rank'
-
--- | Insert an item into the sequence. It may only be drawn from the
--- sequence with 'next' when its counter is next one.
-insert :: (Enum counter, Hashable counter, Eq counter) => Sequential counter a -> a -> STM ()
-insert sequential msg = modifyTVar' (buffer sequential) (HashMap.insert (rank sequential msg) msg)
-
--- | Get and remove the next item from the sequence. This will block until
--- there is an item with the exact rank we are expecting.
-next
-  :: (Enum counter, Eq counter, Hashable counter)
-  => Sequential counter a  -- ^ How item are ordered in sequence
-  -> STM a  -- ^ The next item
-next thingy = do
-  i <- readTVar (current thingy)
-  q <- readTVar (buffer thingy)
-  case HashMap.lookup i q of
-    Nothing -> retry
-    Just msg -> do
-      modifyTVar' (current thingy) succ
-      modifyTVar' (buffer thingy) (HashMap.delete i)
-      pure msg
+receiveMessage session = snd <$> Sequential.next (inbound session)
